requirements.txt
aiofiles==22.1.0
aiogram==3.0.0b5
aiohttp==3.8.4
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
branca==0.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==7.1.2
colorama==0.4.6
comtypes==1.2.0
DAWG-Python==0.7.2
dlib==19.24.2
docopt==0.6.2
exceptiongroup==1.1.1
face-recognition==1.3.0
face-recognition-models==0.3.0
folium==0.14.0
frozenlist==1.3.3
h11==0.14.0
idna==3.4
imutils==0.5.4
Jinja2==3.1.2
magic-filter==1.0.9
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.20.2
openai==0.27.8
opencv-face-recognition==1.0.2
opencv-python==4.5.1.48
outcome==1.2.0
Pillow==8.2.0
PyAudio==0.2.13
pycparser==2.21
pydantic==1.10.9
pyfiglet==0.8.post1
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pyttsx3==2.90
pytz==2023.3
pywin32==306
requests==2.31.0
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SpeechRecognition==3.10.0
srt==3.5.3
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.6.3
urllib3==2.0.3
vosk==0.3.45
Wave==0.0.2
websockets==11.0.3
wsproto==1.2.0
yarl==1.9.2
